==> notebooks/2_Data_Wrangling.ipynb <==
import pandas as pd

megaline_calls=pd.read_csv('../data/raw/megaline_calls.csv')
megaline_messages=pd.read_csv('../data/raw/megaline_messages.csv')
megaline_internet=pd.read_csv('../data/raw/megaline_internet.csv')

# Leer data megaline_users
megaline_users = pd.read_csv('../data/raw/megaline_users.csv')
#Reemplazar los nulos de churn_date por '0000-00-00'
megaline_users['churn_date'] = megaline_users['churn_date'].fillna('2261-12-31')

megaline_users.dtypes

# #Eliminar outliers de megaline_calls 'duration' y mega_internet 'mb_used'
# mean=megaline_calls['duration'].mean()
# std=megaline_calls['duration'].std()
# megaline_calls=megaline_calls[megaline_calls['duration']<mean+2*std]
# megaline_calls=megaline_calls[megaline_calls['duration']>mean-2*std]
# #Eliminar outliers de megaline_internet 'mb_used'
# mean=megaline_internet['mb_used'].mean()
# std=megaline_internet['mb_used'].std()
# megaline_internet=megaline_internet[megaline_internet['mb_used']<mean+2*std]
# megaline_internet=megaline_internet[megaline_internet['mb_used']>mean-2*std]


megaline_plans=pd.read_csv('../data/raw/megaline_plans.csv')
#Cambiar usd_per_gb a usd_per_mb realizando la conversion de GB a MB (1GB son 1000 MB), normalmente usado esta conversion en
#Telefonia movil
megaline_plans['usd_per_gb'] = megaline_plans['usd_per_gb']/1000
megaline_plans.rename(columns={'usd_per_gb':'usd_per_mb'}, inplace=True)
megaline_plans.head()


#Cargar cada dataset a clean
megaline_users.to_csv('../data/clean/megaline_users.csv', index=False)

megaline_calls.to_csv('../data/clean/megaline_calls.csv', index=False)

megaline_messages.to_csv('../data/clean/megaline_messages.csv', index=False)

megaline_internet.to_csv('../data/clean/megaline_internet.csv', index=False)
megaline_plans.to_csv('../data/clean/megaline_plans.csv', index=False)
