==> notebooks/3_Data_Modelling.ipynb <==
import pandas as pd
from itertools import product

Se usará modelado estrella ya que no se tiene tanta información como para denormalizar las dimensiones.

Una dimension. Tabla dim_user_plans

Una tabla de hechos tabla fct_sales (guarda id_usuario,  numero de mensajes, duracion total de llamadas, uso total de internet, date), relaciona las tablas de internet, calls y messages en una sola tabla de gastos del usuario por cada fecha (periodo).

#Cargar todos los datasets de clean
megaline_calls = pd.read_csv('../data/clean/megaline_calls.csv')
megaline_internet = pd.read_csv('../data/clean/megaline_internet.csv')
megaline_messages = pd.read_csv('../data/clean/megaline_messages.csv')
megaline_plans = pd.read_csv('../data/clean/megaline_plans.csv')
megaline_users = pd.read_csv('../data/clean/megaline_users.csv')

#Crear tabla dim_user_plans con un merge de megaline_users y megaline_plans
#a través de la columna plan de megaline_users y plan_name de megaline_plans
dim_user_plans=pd.merge(megaline_users,megaline_plans,left_on='plan',right_on='plan_name',how='left')#se hace un left join para que se mantengan todos los user_id
#aunque no tengan un plan asociado
#Se eliminan la columna plan ya que son redundantes
dim_user_plans.drop(columns=['plan'],inplace=True)
dim_user_plans.head()


#Crear tabla fct_sales
#Se añade la columna user_id, date que recopila todos los user_id y la session_date (internet), call_date (calls) y message_date (messages)
#de los datasets de calls, internet y messages respectivamente sin duplicados. 
#se debe relacionar cada user_id con una fecha en la que uso al menos uno de los servicios.


#Obtener cada user_id de las tablas de calls, internet y messages sin duplicados
usuarios=pd.concat([megaline_calls['user_id'], megaline_internet['user_id'], megaline_messages['user_id']]).drop_duplicates()
#Obtener cada fecha de las tablas de calls, internet y messages sin duplicados
dates=pd.concat([megaline_calls['call_date'], megaline_internet['session_date'], megaline_messages['message_date']]).drop_duplicates()

#Cruzar cada user_id con cada fecha
fct_sales=pd.DataFrame(product(usuarios,dates),columns=['user_id','date'])
print(len(dates))
print(len(fct_sales.loc[fct_sales['user_id']==1000]))


#Para cada user_id y fecha se obtiene el total de minutos, mensajes y mb consumidos
#importante agrupar por user_id y date en el caso de que exista mas de un consumo de servicios en un mismo dia
fct_sales=pd.merge(fct_sales,megaline_calls.groupby(['user_id','call_date'])['duration'].sum().reset_index(),how='left',left_on=['user_id','date'],right_on=['user_id','call_date'])
fct_sales.rename(columns={'duration':'total_minutes'},inplace=True)
fct_sales=pd.merge(fct_sales,megaline_messages.groupby(['user_id','message_date'])['id'].count().reset_index(),how='left',left_on=['user_id','date'],right_on=['user_id','message_date'])
fct_sales.rename(columns={'id':'total_messages'},inplace=True)
fct_sales=pd.merge(fct_sales,megaline_internet.groupby(['user_id','session_date'])['mb_used'].sum().reset_index(),how='left',left_on=['user_id','date'],right_on=['user_id','session_date'])
fct_sales.rename(columns={'mb_used':'total_mb'},inplace=True)
fct_sales.head()

#Quitar las columnas que no se necesitan
fct_sales.drop(columns=['call_date','message_date','session_date'],inplace=True)




#Ver si hay nulos o duplicados en las tabla de fct_sales
print("Tabla fct_sales, nulos")
print(fct_sales.isnull().sum())
print("\nTabla fct_sales, duplicados")
print(fct_sales.duplicated().sum())

#A los nulos de total_minutes, total_messages y total_mb se les asigna 0
fct_sales['total_minutes'].fillna(0, inplace=True)
fct_sales['total_messages'].fillna(0, inplace=True)
fct_sales['total_mb'].fillna(0, inplace=True)
print("\nTabla fct_sales, nulos")
print(fct_sales.isnull().sum())

#Se pasan las nuevas tablas a data_clean
dim_user_plans.to_csv('../data/clean/dim_user_plans.csv', index=False)
fct_sales.to_csv('../data/clean/fct_sales.csv', index=False)
