==> notebooks/4_Feature_Engineering.ipynb <==
import pandas as pd
from sklearn.preprocessing import OrdinalEncoder

Uso de las tablas de clean para feature engineering

#Cargar los datos
dim_user_plans=pd.read_csv("../data/clean/dim_user_plans.csv")
fct_sales=pd.read_csv("../data/clean/fct_sales.csv")



#Añadir la columna month y year
fct_sales['date']=pd.to_datetime(fct_sales['date'])
fct_sales['month']=fct_sales['date'].dt.month
fct_sales['year']=fct_sales['date'].dt.year
#Quitamos la columna date
fct_sales=fct_sales.drop(columns=['date'])
#Agrupar por mes,año y user_id
fct_sales=fct_sales.groupby(['month','year','user_id']).sum().reset_index()
fct_sales.head()


print(fct_sales.loc[fct_sales['month']==12])


#Tabla que 
#Añadir el costo total viendo si un usuario uso más de lo que abastece su plan por cada mes
fct_sales=pd.merge(fct_sales,dim_user_plans,on='user_id',how='left')
fct_sales['exceso_messages'] = fct_sales.apply(lambda row: max(0, row['total_messages'] - row['messages_included']), axis=1)
fct_sales['exceso_mb_used'] = fct_sales.apply(lambda row: max(0, row['total_mb'] - row['mb_per_month_included']), axis=1)
fct_sales['exceso_minutes'] = fct_sales.apply(lambda row: max(0, row['total_minutes'] - row['minutes_included']), axis=1)
fct_sales['costo_extra_messages'] = fct_sales['exceso_messages'] * fct_sales['usd_per_message']
fct_sales['costo_extra_mb_used'] = fct_sales['exceso_mb_used'] * fct_sales['usd_per_mb']
fct_sales['costo_extra_minutes'] = fct_sales['exceso_minutes'] * fct_sales['usd_per_minute']
fct_sales['costo_extra'] = fct_sales['costo_extra_messages'] + fct_sales['costo_extra_mb_used'] + fct_sales['costo_extra_minutes']
fct_sales['total_cost'] = fct_sales['usd_monthly_pay'] + fct_sales['costo_extra']
print(fct_sales.head())


#Se añaden una feature que trabaje con city y tome la primera palabra (antes del espacio xq hay 2 palabras) después de la coma para tener el código de la ciudad
fct_sales['city_code']=fct_sales['city'].str.split(',').str[1].str.split(' ').str[1]

#Usar ordinal encoder para brindar un orden a las ciudades
fct_sales['city_code']=OrdinalEncoder().fit_transform(fct_sales[['city_code']]).astype(int)
fct_sales.head()

fct_sales['reg_date'] = pd.to_datetime(fct_sales['reg_date'])

# Extrae mes y año de 'reg_date'
fct_sales['reg_month'] = fct_sales['reg_date'].dt.month
fct_sales['reg_year'] = fct_sales['reg_date'].dt.year

fct_sales = fct_sales[(fct_sales['reg_year'] <= fct_sales['year'])  & (fct_sales['reg_month'] <= fct_sales['month'])]



# Muestra el resultado
print(fct_sales.head())


#Extraer mes y año de churn_date
fct_sales['churn_date'] = pd.to_datetime(fct_sales['churn_date'])
fct_sales['churn_month'] = fct_sales['churn_date'].dt.month
fct_sales['churn_year'] = fct_sales['churn_date'].dt.year
fct_sales=fct_sales[(fct_sales['churn_year'] >= fct_sales['year'])  & (fct_sales['churn_month'] >= fct_sales['month'])]
print(fct_sales.head())

#Cambiar la columna month y year a otras similares pero con ordinal Encoder
fct_sales['month']=OrdinalEncoder().fit_transform(fct_sales[['month']]).astype(int)
fct_sales['year']=OrdinalEncoder().fit_transform(fct_sales[['year']]).astype(int)

#Cambiar la columna plan_name a otra similar pero con ordinal Encoder
fct_sales['plan_name']=OrdinalEncoder().fit_transform(fct_sales[['plan_name']]).astype(int)

#cargar el one big table para ML a ml
one_big_table=fct_sales.copy()
one_big_table.to_csv('../data/ml/one_big_table.csv',index=False)
