==> notebooks/5_Training_Evaluation.ipynb <==
import pandas as pd
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import SGDRegressor
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
import numpy as np
from scipy.stats import ttest_rel

#Cargar el dataset
df=pd.read_csv("../data/ml/one_big_table.csv")


# Calcular los promedios para cada 'user_id'
df_ml_mean = df.groupby('user_id').agg({
    'costo_extra_messages': 'mean',
    'costo_extra_minutes': 'mean',
    'costo_extra_mb_used': 'mean',
    'costo_extra': 'mean',
    'exceso_messages': 'mean',
    'exceso_minutes': 'mean',
    'exceso_mb_used': 'mean',
    'total_messages': 'mean',
    'total_minutes': 'mean',
    'total_mb': 'mean'
}).reset_index()

# Calcular las desviaciones estándar para cada 'user_id'
df_ml_std = df.groupby('user_id').agg({
    'costo_extra_messages': 'std',
    'costo_extra_minutes': 'std',
    'costo_extra_mb_used': 'std',
    'costo_extra': 'std',
    'exceso_messages': 'std',
    'exceso_minutes': 'std',
    'exceso_mb_used': 'std',
    'total_messages': 'std',
    'total_minutes': 'std',
    'total_mb': 'std'
}).reset_index()

# Unir los resultados de los promedios y desviaciones estándar con el df original
df = df.merge(df_ml_mean, on='user_id', how='left', suffixes=('', '_promedio'))
df = df.merge(df_ml_std, on='user_id', how='left', suffixes=('', '_std'))
df.head()


#Dropear las columnas no utiles para el modelo como user_id
#o que se derivan del precio de los servicios
#como costo_extra_messages, costo_extra_minutes, costo_extra_mb_used, costo_extra
df.drop(columns=["user_id","first_name","last_name",'churn_date','churn_month','churn_year','reg_date','reg_month','reg_year'
                 ,'usd_per_mb','usd_per_minute','usd_per_message','usd_monthly_pay','mb_per_month_included','minutes_included','messages_included',
                 'city','costo_extra_messages','costo_extra_minutes','costo_extra_mb_used','costo_extra',
                    'total_messages','total_minutes','total_mb','exceso_mb_used','exceso_minutes','exceso_messages'],inplace=True)
print(df.columns)


print(len(df))
df.head()



#Valores nulos
df.isnull().sum()


#Reemplazar los nulos por 0
df.fillna(0,inplace=True)

Se asume una distribucion normal por la cantidad de data, por lo que se hace un standardScaler. Sin embargo, solo se escalan datos de costos, etc.

#Conjuntos de entrenamiento y prueba, se quita el indice de pandas

X =df.drop(columns=["total_cost"]).reset_index(drop=True)
y = df["total_cost"]
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.10, random_state=42)
#Eliminar indices
X_train.reset_index(drop=True,inplace=True)
X_test.reset_index(drop=True,inplace=True)
y_train.reset_index(drop=True,inplace=True)
y_test.reset_index(drop=True,inplace=True)
#Escalado de datos
scaler = StandardScaler()
#Se escalaran las ccolumnas total_minutes,total_messages,total_mb_used,exceso_messages,exceso_minutes,exceso_mb_used
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)
#Convertir a arreglos numpy y_train y y_test
y_train = y_train.values
y_test = y_test.values
print(X_train)
print(X_test)
print(y_train)
print(y_test)

def obtener_scores(modelo, X_train, y_train, X_test, y_test):
    y_pred_train = modelo.predict(X_train)
    y_pred_test = modelo.predict(X_test)
    rmse_train= np.sqrt(mean_squared_error(y_train, y_pred_train))
    rmse_test= np.sqrt(mean_squared_error(y_test, y_pred_test))
    mae_train= mean_absolute_error(y_train, y_pred_train)
    mae_test= mean_absolute_error(y_test, y_pred_test)
    r2_train= r2_score(y_train, y_pred_train)
    r2_test= r2_score(y_test, y_pred_test)
    salida= {"RMSE_train":rmse_train,"MAE_train":mae_train,"R2_train":r2_train,"RMSE_test":rmse_test,"MAE_test":mae_test,"R2_test":r2_test}
    return salida

#Regresión lineal usando scikit-learn LinearRegression
model = LinearRegression()
model.fit(X_train, y_train)
y_pred = model.predict(X_test)
#Scores
print(obtener_scores(model, X_train, y_train, X_test, y_test))



#Regresión lineal con SGDRegressor que use regularización L1 y K-Fold Cross Validation para hallar el mejor valor de alpha
# y learning rate (eta0)
from sklearn.model_selection import GridSearchCV
params = {'alpha': [0.0001, 0.001, 0.01, 0.1, 1, 10, 100],
          'eta0': [0.0001, 0.001, 0.01, 0.1, 1, 10, 100]}
sgd = SGDRegressor(penalty='l1', max_iter=1000)
grid = GridSearchCV(sgd, params, cv=5)
grid.fit(X_train, y_train)
print(grid.best_params_)
model = grid.best_estimator_
model.fit(X_train, y_train)
#Scores
obtener_scores(model, X_train, y_train, X_test, y_test)



#Regresión polinomial usando PolynomialFeatures, Stochastic Gradient Descent, elastic net
# y K-Fold Cross Validation para hallar el mejor valor de alpha y el mejor grado del polinomio
#Para todo esto también se debe usar un pipeline para encadenar los procesos e incluir PolynomialFeatures en el modelo
from sklearn.preprocessing import PolynomialFeatures
from sklearn.pipeline import make_pipeline
params = {'polynomialfeatures__degree': [1, 2, 3], 'sgdregressor__alpha': [0.0001, 0.001, 0.01, 0.1, 1, 10],
          'sgdregressor__eta0': [0.01, 0.1, 1, 10]}
model= make_pipeline(PolynomialFeatures(), SGDRegressor(penalty='elasticnet', max_iter=100))
grid = GridSearchCV(model, params, cv=2,verbose=1)
grid.fit(X_train, y_train)
print(grid.best_params_)
model = grid.best_estimator_
model.fit(X_train, y_train)
#Scores
obtener_scores(model, X_train, y_train, X_test, y_test)

# Ajustar el modelo de LinearRegression
model1 = LinearRegression()
model1.fit(X_train, y_train)
y_pred1 = model1.predict(X_test)
print(obtener_scores(model1, X_train, y_train, X_test, y_test))
# Ajustar el modelo de PolynomialFeatures con SGDRegressor
model2 = make_pipeline(PolynomialFeatures(degree=1), SGDRegressor(penalty='elasticnet', max_iter=10000, alpha=0.1, eta0=0.01))
model2.fit(X_train, y_train)
y_pred2 = model2.predict(X_test)
print(obtener_scores(model2, X_train, y_train, X_test, y_test))

# Realizar la prueba t de hipótesis: Evaluar la diferencia entre los modelos
# Hipótesis nula: modelo1 es mejor o igual que modelo2 (si y_pred1 >= y_pred2)
# Comparar las diferencias entre las predicciones de ambos modelos
differences = y_pred1 - y_pred2

# Realizar la prueba t de hipótesis sobre las diferencias
t_stat, p_value = ttest_rel(differences, [0] * len(differences))  # Comparando las diferencias con 0

# Imprimir el resultado de la prueba t
print(f"Estadístico t: {t_stat}")
print(f"Valor p: {p_value}")

# Conclusión según el valor p
if p_value < 0.05:
    print("Rechazamos la hipótesis nula: el modelo con PolynomialFeatures y SGDRegressor es significativamente mejor.")
else:
    print("No rechazamos la hipótesis nula: no hay evidencia suficiente para decir que el modelo con PolynomialFeatures y SGDRegressor es mejor.")


#Score final
obtener_scores(model1, X_train, y_train, X_test, y_test)
